+++ src/main/php/auth/AuthenticationProvider.php
@@ -31,10 +31,24 @@ interface AuthenticationProvider
     /**
      * returns login uri
      *
-     * The method is called when the authenticate() method returns <null>.
+     * The method is called when the authenticate() method returns <null> and a
+     * redirect to a login URI is allowed for the resource.
      *
      * @param   \stubbles\webapp\Request  $request
      * @return  string|\stubbles\peer\http\HttpUri
      */
     public function loginUri(Request $request);
+
+    /**
+     * returns a list of challenges to send in response's 401 WWW-Authenticate header for given request
+     *
+     * The method is called when the authenticate() method returns <null> and a
+     * redirect to a login URI is not allowed for the resource, but a
+     * 401 Unauthorized response should be send instead.
+     *
+     * @since   8.0.0
+     * @param   \stubbles\webapp\Request  $request
+     * @return  string[]  list of challenges for the WWW-Authenticate header, must at least contain one
+     */
+    public function challengesFor(Request $request): array;
 }